Adding audio support
Cuttlefish on Android-12 is using the audio HAL from goldfish. The audio HAL for goldfish doesn't support AUDIO_DEVICE_OUT_BUS as an audio device. Android Automotive variant needs the primary output device to be AUDIO_DEVICE_OUT_BUS , which is what is used in the Cuttlefish Auto on Android-12.
So enabling the AUDIO_DEVICE_OUT_BUS on goldfish audio HAL.